+++ clarite/tests/r_test_output/analyze/nhanes_average_result.csv
