--- clarite/tests/r_test_output/nhanes_complete_result_subset_cat.csv
@@ -1,4 +1,0 @@
-"Variable","Variable_type","N","Beta","SE","Diff_AIC","pvalue"
-"race","categorical",5941,NA,NA,40.8452346434315,0.245877287540186
-"agecat","categorical",5941,NA,NA,-315.477184355315,9.88256654421118e-08
-"RIAGENDR","binary",5941,0.38903867405612,0.106573017991987,NA,0.00445903421762822
