--- clarite/tests/r_test_output/nhanes_complete_result_subset_cont.csv
@@ -1,4 +1,0 @@
-"Variable","Variable_type","N","Beta","SE","Diff_AIC","pvalue"
-"race","categorical",3955,NA,NA,25.2873498940394,0.271831835407314
-"agecat","categorical",3955,NA,NA,-138.31900523803,7.59650343615736e-05
-"RIAGENDR","binary",3955,0.222803552426531,0.123790703497555,NA,0.105417238490215
