--- clarite/tests/r_test_output/nhanes_complete_withna_result.csv
@@ -1,4 +1,0 @@
-"Variable","Variable_type","N","Beta","SE","Diff_AIC","pvalue"
-"race","categorical",7113,NA,NA,49.5642746451613,0.235356961224573
-"agecat","categorical",7113,NA,NA,-298.047690965284,3.86385190534053e-06
-"RIAGENDR","binary",7113,0.157317961657818,0.0862154198306354,NA,0.101343421974207
