--- clarite/tests/r_test_output/nhanes_weightsonly_result.csv
@@ -1,4 +1,0 @@
-"Variable","Variable_type","N","Beta","SE","Diff_AIC","pvalue"
-"race","categorical",7846,NA,NA,50.9701263969109,0.0668433839624171
-"agecat","categorical",7846,NA,NA,-302.875821281627,1.28762477627071e-48
-"RIAGENDR","binary",7846,0.212760495203168,0.0970296204313762,NA,0.0283555792758872
